Use unique names for each figure
Recent changes in GMT add the option of going back to a previous figure
using its prefix. We were using the same prefix for all figures so we're
always plotting on the same one. Use NamedTemporaryFile to generate a
unique name and pass that as prefix.

Updated some of the baselines that changed slightly.