Refactor clib funcs to methods of the context manager (#42)
The functions in `gmt.clib` need to keep be passed the session pointer 
and users have to remember to get it from `APISession`, 
which just handles creating and deleting the pointer. 

For example:

    with APISession() as session:
        call_module(session, ...)

Turn the functions into methods of the context manager and rename it `LibGMT`. 
Hide away the session management functions as private methods.
So now it looks like:

    with LibGMT() as lib:
        lib.call_module(...)

Fixes #38