C lib access through funcs called by LibGMT
The class LibGMT will start getting big with the wrappers for creating
and accessing data. Move the actual wrapper code to functions in
gmt.clib.core (like it was before) and have LibGMT methods call these
functions (passing in the loaded library and session). This will allow
us to separate the functions into files if we need it without breaking
the class.