Lowlevel wrapper for GMT_Create_Data (#44)
Implement a low-level hook to `GMT_Create_Data` (used to create 
a GMT dataset that can be populated with numpy array data).

Introduces the new module `gmt/clib/io.py`. The function is not 
accessible from the public API (i.e., `LibGMT`). They will be serve as 
a basis for a high-level virtual file API.

Lays the foundation for #22