Wrap GMT_Put_Vector and GMT_Write_Data (#74)
Used to insert numpy 1d array into a GMT_Dataset and to write the dataset 
to a file so we can test it. 
Limit write_data to output to a file only (no streams). That would be too 
troublesome going from Python to C and not that useful anyway.
The vectors need to be C contiguous when passed to put_vector. 
Can't change them inside the method because the array will be garbage 
collected when the method returns and the memory will be wiped. 
This must be done by the function calling create_data and put_vector.