Tutorials, download ipynb, edit on github
Move first steps into a tutorials folder that will hold other notebooks.
Replace "View source" with "Fount a typo?" which goes to edit the file
on github or open an issue with the typo.
Add download button for notebook generated pages. Downloads from Github.
Include link to the online demo on mybinder.