Move helper functions/classes/etc into gmt.helpers
Create a new package called gmt.helpers that contains all of the helper
functions, decorators, classes, and context managers. Helps organize the
imports and establish an API for the helpers. Include a listing in the
API docs for this package. This way, the main modules will contain only
the actual GMT wrapper code.