Add preview option using NASA WorldWind Web
Use the javascript virtual globe to display a KML preview of the figure
in the notebook.
The preview files (KML and PNG) need to be in the same directory as the
notebook for things to work. The /tmp dir is outside of the range of the
notebook server. So they need to be deleted separately.
Include an option to not delete the preview files so that the globe
works when the notebook is converted to HTML (for the docs and
nbviewer).
Need to include require.js in order to load the javascript library in
the documentation the same way as is done in the notebook.
Not extensively tested but the notebook works.