Use base64 png instead of kml in worldwind preview
Was using a KML preview loaded on WorldWind Web for the globe preview.
The problem was that the file needed to be in the same folder as the
notebook and couldn't be deleted after the preview was used.
Got around that by using a SurfaceImage layer instead of a KML layer.
The SurfaceImage takes a PNG image and region and displays it on the
globe.
The PNG can be inserted as a base64 encoded string using the Javascript
Image object.
This eliminates the need to keep the preview files around.

I capture the plot region by saving a KML preview and extracting the
bounding box from the KML file.
This is better than intercepting the 'region' argument in the
_preprocess method because it works for ISO country codes as well.
With the region, I can automatically set the view point.