Add function to load Earth relief data (#156)
Implements the function `load_earth_relief` in `gmt.datasets` to load
the Earth relief grids distributed with GMT (the `@earth_relief_XXm` 
special file names). Uses `gmt.which` to download to the user directory 
and get the file name. The netCDF grid is loaded and returned as an 
xarray.DataArray. Only support the 60 to 1 arc-minute global grids. 
Higher resolution arc-second grids are served as tiles and are usually
cut to a specific area, so a `region` argument will be needed to load
these grids later, probably in conjunction with `gmt.grdcut`.