Using LibGMT in 'with' fails if GMT < 6.0.0 (#175)
Uses the new LibGMT.info to check if GMT is at least 6.0.0. Raises a
GMTVersionError exception if it's not.

It would be great if we could check this upon instantiation but that
wouldn't work because `LibGMT.get_default` requires an active session.
So the only logical place to put this is in `LibGMT.__enter__` after the
session is created. The session is closed before the exception is
raised.

Fixes #171