Set libgmt path with GMT_LIBRARY_PATH from environ (#177)
LibGMT no longer takes a pathname as input. The library path is either
just the file name, or the directory is taken from the GMT_LIBRARY_PATH
environment variable. This allows users to direct the GMT/Python to
libgmt. It's useful to test against a local build of GMT trunk or
resolving path issues without resorting to LD_LIBRARY_PATH.

Replace section on LD_LIBRARY_PATH with GMT_LIBRARY_PATH in install
docs. Make it clear that this should be a last resort.

Fixes #172 and fixes #165