Remove work done in LibGMT.__init__
The shared library was being loaded in __init__, which is not a good
design pattern. Instead, now it's loaded the first time it gets used by
the new method `get_libgmt_func` that gets the ctypes function and sets
the return and argument types. This removes some redundancy from the
other methods (setting ctypes type conversions) and also removes the
usage of a private variable `_libgmt` from other methods. Now, only
methods that set a private variable need to know of it's existence.
The same applies to `_session_id` which is entirely encapsulated by
`current_session`.