Log error messages with a callback instead of file
Use a callback function passed to `GMT_Create_Session` to log error
messages instead of redirecting them to a file using
`GMT_Handle_Messages`. It's a lot cleaner and the code is more legible.
Other functions don't need to do anything to have their errors logged,
it's automatic.
The logged messaged are also printed to stderr so they will show up in
the Jupyter notebook.
This is useful when using the verbose mode (`V="d"`) in modules.