Log error messages with a callback instead of file (#188)
Use a callback function passed to `GMT_Create_Session` to log error
messages instead of redirecting them to a file using
`GMT_Handle_Messages`. It's a lot cleaner and the code is more legible.
Other functions don't need to do anything to have their errors logged,
it's automatic.
The logged messaged are also printed to stderr so they will show up in
the Jupyter notebook.
This is useful when using the verbose mode (`V="d"`) in modules.

Switch to the Fatiando a Terra CI scripts and enable OSX testing on 
Travis.
Fatiando provides scripts for handling all of the CI tasks we need: 
https://github.com/fatiando/continuous-integration 
Use them instead of rewriting everything.

Fixes #164 
The Segmentation fault on OSX was happening because of the log file
that we used to capture the GMT output and include in the exception.
I have no idea why this happens.
But removing that fixes the issue so I'm happy not knowing.