Sphinx extension for GMT/Python plots (#208)
Add a sphinx extension `gmt.sphinxext.gmtplot` to the package.
It enables the `gmt-plot` directive that executes the given code,
captures the output of `gmt.Figure.show` and stdout, and 
inserts them into the rst document.

Includes a plot in the front page and a page documenting the
extension. It also serves as a test for the extension.

The code is not tested with pytest. This could possibly be done
using [sphinx-testing](https://github.com/sphinx-doc/sphinx-testing).

Fixes #129