Make the API reference manually (#209)
The automatically generated one is a bit difficult to follow and there
is no separation between different types of functions/classes.
The Figure and LibGMT methods are listed on the main page instead of
only in the class pages. Each method has its own page, which will make
it easier to include an example plot and when the library grows.