Updated `gmt.clib.Session.__getitem__` to reflect updated GMT 6.0.0 API for `GMT_Get_Enum`. Resolves #222
This update breaks compatability with trunk revisions prior to r20437. However,
a rudimentary check is done to try to indicate to the user. It can be removed
once GMT 6.0.0 is released.

Additionally, session handling has been added to `__getitem__` to use the
session pointer if there is an open session.

Warning added to API documentation