Drop Py35, update guides, Makefile, and CI config
Copied over some updated guides and CI configuration from Fatiando a
Terra. Includes a maintenance guide for review PRs and making releases.
Some Makefile rules have changed to make things simpler.

Also decided to drop Python 3.5 so we don't have to test against it or
be limited by things that are 3.6+ only. Should be OK since most people
are on 2.7 or 3.6 and Anaconda now ships with 3.7 by default.

Simplified the README as well because PyPI will fail to render the `raw`
rst directive.

Fixes some lint errors because of an update in pylint.