Build GMT master instead of using conda-forge (#261)
The conda-forge development package creates all sorts of problems for us with 
mixtures of different channels, etc. It's a hassle to setup and we end up falling behind
new features added to GMT 6 (which is what we support).

Replace the conda-forge package with GMT built from the master branch on the CI
services. Update the install instructions to tell users to build GMT from source.