Rename the package to PyGMT (#265)
GMT/Python is a bad name because it can't be translated into a
repository or Python package name easily. What we have now is a
repository called `gmt-python`, a website `gmtpython`, a package `gmt`,
and a PyPI package `gmt-python`. This is madness.
Since all other GMT Python wrappers are dead, we'll rebrand as PyGMT
(repo and packages will be `pygmt`).

This renames the package and imports so that the tests pass and the docs
build. I'm sure we'll find many references to `gmt` but we can fix those
with time.

Fixes #263