Use sphinx-gallery to manage examples and tutorials (#268)
Enables the usage of sphinx-gallery with PyGMT. For now, requires features that aren't
yet released so we need to install from the Github master branch of sphinx-gallery.
Implements an image scrapper for `pygmt.Figure` that captures them and saves to the
correct file names. Need to add a global figure registry (`SHOWED_FIGURES` in
`pygmt/figure.py`) so that we can request all figure objects that have called their
`show` method. Replace the current jupyter notebooks by gallery notebook style `.py`
files, which are easier to maintain on version control.