Assign authorship to "The PyGMT Developers"
Replaces my name with "The PyGMT Developers" in authorship claims and
copyright notices. Include a definition of what that term means in the
AUTHORS.md file and include a reference to it in the README and doc
page.
Left my email on the `setup.py` because we don't have a project email
yet.