Add Google Analytics to measure site visitors
Add the tracking script to the documentation template so we can know how
many people are visiting the docs and which pages they see the most.
The data are anonymous and we do not store cookies to track users in
other sites. This limits the amount of information we can get but is
much more ethical and less creepy.