Refresh installation documentation links and use conda-forge again
Bringing back the conda-forge way of installing GMT binaries as we anticipate the release of GMT 6. Note though that PyGMT still isn't tested to work on Windows. Switched to using `conda activate` instead of `source activate` which is cross-platform and available for conda>=4.6. Also updated some http links to use https if possible, pointing some anaconda ones to new locations.