Enable multiple textfile input to text
Using the sequence_space processor in #325 to allow for passing in a list of textfiles to be plotted! Problem with that `text` would consider a numpy matrix as a list to be parsed, so the code was updated to fix that.

Unit tests and docstrings updated. Strangely enough, the 'projection'/'-J' argument doesn't seem to be required anymore? Also regenerated baseline plots to use their rightful name...