Improve error handling for failing to load dll properly
Reverse the list of libnames for Windows so that 'gmt.dll' (the one in our conda build) is returned instead of 'gmt_w32.dll' when things go wrong. Also raise a more helpful error message with the aid of ctypes.util.find_library, and fix logic to reset the 'error' variable to False once library loads successfully.