Fix how DLLs are found
Here we need to set `error` to True first, so if libnames is empty, than
we raise the GMTCLibNotFoundError error.

On Windows, libnames is a list of possible DLL names, "gmt.dll", "gmt_w64.dll", and "gmt_w32.dll".
If any one of them is found, then we set `error` to False and break to for loop.

If none of them are found, then we raise the GMTCLibNotFoundError error:
```
Error loading the GMT shared library 'gmt.dll, gmt_w64.dll, gmt_w32.dll'.
```

The error message above can be improved, but it's not a big issue.