Change load_earth_relief()'s default resolution to 01d (#488)
In GMT 6.0.0, both `01d` and `60m` are valid resolutions of earth
relief data. It was called `60m` at the beginning and was changed to `01d`
when GMT 6.0.0 was officially released. `60m` is still valid for backward
compatibility.

Run the following commands, and you will have the two data in the
current directory.
```
gmt which @earth_relief_60m -Gl
gmt which @earth_relief_01d -Gl
```
These two files have different file names but are identical:
```
$ md5sum earth_relief_01d.grd earth_relief_60m.grd
74a884c902015dda516d17605f317efe  earth_relief_01d.grd
74a884c902015dda516d17605f317efe  earth_relief_60m.grd
```

In the upcoming GMT 6.1.0, the resolution `60m` will be deprecated.
That's why we have many ~25 errors when testing PyGMT with the GMT master
branch, simply because GMT 6.1.0 can't download the `@earth_relief_60m`.

To make the transition to GMT 6.1.0 easier, here I change the default
earth relief resolution of `load_earth_relief()` function from `60m` to
`01d`. As the two grids are identical, the change in this PR won't break
anything.

Note that, currently there are ~43 failures due to the recent updates of
the GMT data server (#451), and we can't fix these failures easily due
to the grid registration issue (#476). Thus, I don't try to fix any
failures in this PR.

The test log files of the master branch and this branch are the same.
Tests that fail in the master branch still fail in the same way in this branch.

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>