Improve the cache step in GitHub Action
We use the `wget` command to download GMT remote and cache data, because
`gmt which` command always timeouts in Windows and Linux jobs, but work
well in macOS jobs.

It turns out it's a problem of the GMT official (oceania) data server.
The europe data server work well on all platforms. Thus, we can use
`gmt which` on all platforms.

Changes in this PR:

- Use the europe data server: https://fct-gmt.ualg.pt/gmt/data
- Remove the workaround for Linux and Windows
- Update the date in the cache key