Fix a failing test due to GMT API change
The test `test_put_matrix_grid` fails due to recent changes of GMT API
in https://github.com/GenericMappingTools/gmt/pull/3848.

Before this PR, for a matrix grid, calling GMT_Write_Data always writes
the matrix grid as a matrix, no matter the geometry is `GMT_IS_POINT` or
`GMT_IS_SURFACE`.

This PR declares it as a bug (or an improvement). With that PR merged,
`geometry=GMT_IS_POINT` writes a matrix, `geometry=GMT_IS_SURFACE`
writes a the matrix grid as a netCDF file.

This test expects a matrix, but the new API writes a netCDF grid.
That's why it fails.

This PR fixes the issue by simply changing `GMT_IS_SURFACE` to
`GMT_IS_POINT`, so that the test can pass for GMT 6.1.0, 6.1 and master
branches.