Sensible array outputs for pygmt info
When either of per_column (C), spacing (I), or
nearest_multiple (T) are used in `pygmt.info`,
output the result as a numpy.ndarray which
would be more usable that a raw string that is
meant for the command line world. Also improve
the docstring of `pygmt.info` to mention that
numpy.ndarray outputs are being reported.