Improve how PyGMT find the GMT library
In the old codes, PyGMT first checks if the environmental variable `GMT_LIBRARY_PATH`
is defined. If yes, it tries to find the GMT library in the specified path.
Otherwise, it will search for the GMT library in standard library paths
(LD_LIBRARY_PATH on Linux; DYLD_LIBRARY_PATH on macOS; PATH on Windows).

This PR improve how PyGMT find the library, by search for all possible
paths. The paths from the highest priority to the lowest priority are:

1. the path defined by `GMT_LIBRARY_PATH`
2. the path returned by command `gmt --show-library`
3. On Windows, also check the path returned by `find_library` function
4. the standard system paths

Thus, the function `clib_full_names` now returns a list of possible
names, for example:
```
["/path/defined/by/gmtlibrarypath/libgmt.so",
 "/usr/local/lib/libgmt.so",
 "libgmt.so"]
```

**Pros**:

1. More ways to find the library, so less "GMTCLibNotFoundError" errors

**Cons**:

1. It calls the "gmt" command in a new process
2. A much longer list returned by `clib_full_names`. Maybe slower loading?
3. Difficult to test.

Closes #628.