Fix the error message when IPython is not available
When IPython is not available, `fig.show()` reports an error:
```
Cannot find IPython. Make sure you have it installed or use 'external=True' to open in an external viewer.
```

The error message is out-of-date. Instead, it should report:
```
Cannot find IPython. Make sure you have it installed or use 'method="external"' to open in an external viewer.
```