Let GMT skip refreshing the server data files (#866)
The GMT caches and remote data are cached in GitHub Actions by the
cache_data.yaml workflow, because the Linux and Windows agents can't
connect to the GMT's data server. The "Cache data" workflow is run weekly.

When GMT needs to access a remote file, it first checks if ~/.gmt/server/gmt_data_server.txt
is older than 1 day. If yes, then GMT downloads the newer version of the file
from the GMT data server. As the Linux and Windwos agents have network connection
issues, the tests are slower due to the internet connection.

In this PR, we run the touch command to update the modification times of the
GMT's two server files, so that GMT won't try to refresh it.