Open images using the associated application on Windows
This PR improves the `launch_external_viewer` function to open images
using the associated application (just like double-click the images) on
Windows.

The implemention is done by calling the
[`os.startfile`](https://docs.python.org/3/library/os.html#os.startfile)
function.

`os.startfile` is only available on Windows, so need to disable the pylint
error `no-member`.

The code was originall written in #269 by @leouieda, re-used in #529.