CI: Do NOT fail CI if codecov runs into errors during upload
The codecov github action is very unstable recently, and make most of
our CI jobs fail.

This PR sets `fail_ci_if_error` to `false` (the default behavior of
codecov), so that even if codecov fails in uploading the coverage
reports, the CI job still pass.

Actually, if the codecov upload fails, we will see the failing codecov
checks, but at least we know that all tests pass.