Add a gallery example showing individual custom symbols
So far we have no gallery example in which custom symbols are shown. This PR adds such an example to the gallery. Since the number of custom symbols coming with GMT is huge (https://docs.generic-mapping-tools.org/latest/cookbook/custom-symbols.html) I think we should focus only on a subset of them.