Wrap grdmath
Initial commit for wrapping the grdmath function for #916
which is a "Raster calculator for grids (element by element)".
Original GMT `grdmath` documentation is at
https://docs.generic-mapping-tools.org/6.2/grdmath.html.
Implementation works by building and storing a computational
graph of the grid operations in a GrdMathCalc class object.
An output NetCDF grid or xarray.DataArray is produced
upon calling `.compute()` or by setting `outgrid=True`.