Add Mars dataset (#1420)
Sample data table for the shape of Mars. This is the
`@mars370d.txt` dataset used in GMT examples, with data
and information from Smith, D. E., and M. T. Zuber (1996).
The shape of Mars and the topographic signature of the
hemispheric dichotomy. Data columns are "longitude",
"latitude", and "radius (meters)."

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>