Update text symbol gallery example
According to a recent [forum post](https://forum.generic-mapping-tools.org/t/changing-symbol-outline-color-using-plot/2410), here's an update for the text symbol gallery example. Since it is not fully clear from the current example how to change the outline color of the symbols, for the last symbol now the outline color is changed.