Add Figure.tilemap to plot XYZ tile maps (#2394)
Adding the tilemap method for plotting XYZ tile maps.
This is a wrapper around `pygmt.datasets.load_tile_map`
and `pygmt.Figure.grdimage`. Aliases from `grdimage` have
been copied over, and docstring for parameters from
`load_tile_map` have been copied over too.

* Add rioxarray to CI build matrix and include it as optional dependency

Let the Continuous Integration tests run with `rioxarray`,
include it in pyproject.toml and environment.yml, and document
it in `doc/install.rst` as an optional dependency.

* Reproject and plot map in lonlat coordinates by default

Given a region in lonlat coordinates, ensure that the output
figure is also plotted in lonlat instead of Web Mercator.

* Ensure that plot is clipped to bounding box region when no_clip is False

Pass the region to the grdimage call so that the plot extent is
the same as the bounding box region. Set `no_clip=True` to prevent
the clipping from happening (i.e., the plot will extend out from
the region of interest). Added a unit test checking results for
both no_clip True/False, and updated some previous baseline images
that have changed slightly.

* Web Mercator to Spherical Mercator and remove verbose statement

* Use rio.set_crs instead of rio.write_crs

So that the `spatial_ref` CF coordinate won't be set, which would
result in an int32 variable on Windows but int64 on Linux/macOS.

---------

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
Co-authored-by: Yvonne Fröhlich <94163266+yvonnefroehlich@users.noreply.github.com>